made main rendering layer
