+++ src/core/Application.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "LayerStack.h"
 #include "Models/Cube.h"
 #include "Models/Cursor3D.h"
 #include "Models/Prism.h"
@@ -23,17 +24,18 @@ namespace GoL {
 class Application {
 public:
 private:
-    Window window;
-    Camera camera;
-
     bool isRunning = true;
+    Window window;
+    LayerStack layerStack;
 
-    bool firstMouse = true;
-    float lastX = 0, lastY = 0;
+    // Camera camera;
 
-    float currentFrameTime = 0;
-    float lastFrameTime = 0;
-    float deltaTime = 0;
+    // bool firstMouse = true;
+    // float lastX = 0, lastY = 0;
+
+    // float currentFrameTime = 0;
+    // float lastFrameTime = 0;
+    // float deltaTime = 0;
 
 public:
     Application(
@@ -42,7 +44,8 @@ public:
             unsigned int height = 1080
     )
         : window(Window::Data(title, width, height))
-        , camera({ 0.0f, 0.0f, -3.0f }, { 0.0f, 1.0f, 0.0f }, 1920.0f, 1080.0f) {
+    // , camera({ 0.0f, 0.0f, -3.0f }, { 0.0f, 1.0f, 0.0f }, 1920.0f, 1080.0f) {
+    {
         this->window.SetEventCallback(
                 [this](Event& e) {
                     this->OnEvent(e);
@@ -51,37 +54,17 @@ public:
         this->window.Configure();
     }
 
+    void PushLayer(Layer* layer) {
+        this->layerStack.PushLayer(layer);
+        layer->OnAttach();
+    }
+
     void Run() {
-        Renderer renderer;
-        Cursor3DRenderer cursor_renderer;
-
-        Shader shader = { "resources/shaders/test.shader" };
-        shader.Bind();
-
-        auto cube = Cube(glm::vec3 { 0.0f }, glm::vec3 { 0.0f }, 0.5f);
-        cube.BindIndices();
-
-        auto cursor = Cursor3D();
-        cursor.BindIndices();
-
-        auto prism = Prism(glm::vec3 { 1.0f, 1.0f, 0.0f }, glm::vec3 { 0.0f, 0.0f, 0.0f }, 0.5f);
-        prism.BindIndices();
-
-        IndexBuffer::Init();
-
         while (isRunning) {
-            currentFrameTime = static_cast<float>(glfwGetTime());
-            deltaTime = currentFrameTime - lastFrameTime;
-            lastFrameTime = currentFrameTime;
 
-            renderer.Clear();
-
-            renderer.Draw<Cube>(cube, camera, shader);
-            // renderer.Draw<Prism>(prism, camera, shader);
-            cursor.SetScaleFactor(100);
-            renderer.Draw<Cursor3D>(cursor, camera, shader);
-            cursor.SetScaleFactor(0.1);
-            cursor_renderer.Draw<Cursor3D>(cursor, camera, shader);
+            for (auto it = layerStack.begin(); it != layerStack.end(); it++) {
+                (*it)->OnUpdate();
+            }
 
             window.OnUpdate();
         }
@@ -96,76 +79,23 @@ public:
 
         // probably make it async in the future ?
         dispatcher.Dispatch<WindowCloseEvent>(BIND_MEMBER_EVENT_FN(Application::OnWindowClose));
-        dispatcher.Dispatch<KeyPressedEvent>(BIND_MEMBER_EVENT_FN(Application::OnKeyPress));
-        dispatcher.Dispatch<MouseMovedEvent>(BIND_MEMBER_EVENT_FN(Application::OnMouseMove));
+
+        for (auto it = layerStack.rbegin(); it != layerStack.rend(); it++) {
+            if (e.Handled) {
+                break;
+            }
+
+            (*it)->OnEvent(e);
+        }
+
+        // TODO
+        // dispatcher.Dispatch<KeyPressedEvent>(BIND_MEMBER_EVENT_FN(Application::OnKeyPress));
+        // dispatcher.Dispatch<MouseMovedEvent>(BIND_MEMBER_EVENT_FN(Application::OnMouseMove));
     }
 
 private:
     bool OnWindowClose(WindowCloseEvent& e) {
         isRunning = false;
-        return true;
-    }
-
-    bool OnKeyPress(KeyPressedEvent& e) {
-        float cameraSpeed = static_cast<float>(2.5 * deltaTime);
-        if (e.GetKeyCode() == GLFW_KEY_W) {
-            camera.ProcessKeyboard(GoL::CameraMovement::FORWARD, deltaTime);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_S) {
-            camera.ProcessKeyboard(GoL::CameraMovement::BACKWARD, deltaTime);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_A) {
-            camera.ProcessKeyboard(GoL::CameraMovement::LEFT, deltaTime);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_D) {
-            camera.ProcessKeyboard(GoL::CameraMovement::RIGHT, deltaTime);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_LEFT_SHIFT) {
-            camera.ProcessKeyboard(GoL::CameraMovement::UP, deltaTime);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_LEFT_CONTROL) {
-            camera.ProcessKeyboard(GoL::CameraMovement::DOWN, deltaTime);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_Q) {
-            camera.ProcessMouseMovement(-10.0f, 0.0f);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_E) {
-            camera.ProcessMouseMovement(10.0f, 0.0f);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_LEFT) {
-            camera.ProcessMouseMovement(-10.0f, 0.0f);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_RIGHT) {
-            camera.ProcessMouseMovement(10.0f, 0.0f);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_UP) {
-            camera.ProcessMouseMovement(0.0f, 10.0f);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_DOWN) {
-            camera.ProcessMouseMovement(0.0f, -10.0f);
-        }
-        if (e.GetKeyCode() == GLFW_KEY_ESCAPE) {
-            glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_NORMAL);
-            glfwSetCursorPosCallback(window, NULL);
-            firstMouse = false;
-        }
-
-        return true;
-    }
-
-    bool OnMouseMove(MouseMovedEvent& e) {
-        if (firstMouse) {
-            lastX = e.GetX();
-            lastY = e.GetY();
-            firstMouse = false;
-        }
-
-        glm::vec2 offset(e.GetX() - lastX, lastY - e.GetY());
-        lastX = e.GetX();
-        lastY = e.GetY();
-
-        camera.ProcessMouseMovement(offset.x, offset.y);
-
         return true;
     }
 };
